LCE state as a separate artefact References: #44
