+++ MIGRATION_5.md
@@ -1,0 +1,8 @@
+# Migration to version 5.x
+
+# Imports
+
+| Artefact                            | Old package                 | New package                |
+| ----------------------------------- | --------------------------- | -------------------------- |
+| LceState and extensions             | com.motorro.rxlcemodel.base | com.motorro.rxlcemodel.lce |
+| LceUseCase, LceModel and extensions | com.motorro.rxlcemodel.base | com.motorro.rxlcemodel.rx  |
