+++ lce/src/main/kotlin/com/motorro/rxlcemodel/lce/LceState.kt
@@ -1,5 +1,5 @@
 /*
- * Copyright 2019 Nikolai Kotchetkov.
+ * Copyright 2022 Nikolai Kotchetkov.
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
  * You may obtain a copy of the License at
@@ -11,7 +11,7 @@
  * limitations under the License.
  */
 
-package com.motorro.rxlcemodel.rx
+package com.motorro.rxlcemodel.lce
 
 /**
  * State for "Loading-Content-Error" resource which retrieves [data]
