+++ rx/src/main/kotlin/com/motorro/rxlcemodel/rx/CacheThenNetLceModel.kt
@@ -13,7 +13,8 @@
 
 package com.motorro.rxlcemodel.rx
 
-import com.motorro.rxlcemodel.rx.LceState.*
+import com.motorro.rxlcemodel.lce.LceState
+import com.motorro.rxlcemodel.lce.LceState.*
 import com.motorro.rxlcemodel.rx.LogLevel.INFO
 import com.motorro.rxlcemodel.rx.entity.Entity
 import com.motorro.rxlcemodel.rx.service.ServiceSet
