+++ rx/src/main/kotlin/com/motorro/rxlcemodel/rx/lceUtils.kt
@@ -13,6 +13,9 @@
 
 package com.motorro.rxlcemodel.rx
 
+import com.motorro.rxlcemodel.lce.LceState
+import com.motorro.rxlcemodel.lce.combine
+import com.motorro.rxlcemodel.lce.map
 import io.reactivex.rxjava3.core.Completable
 import io.reactivex.rxjava3.core.Observable
 import io.reactivex.rxjava3.core.Single
