+++ rx/src/test/kotlin/com/motorro/rxlcemodel/rx/UpdatingLceModelWrapperTest.kt
@@ -1,5 +1,5 @@
 /*
- * Copyright 2019 Nikolai Kotchetkov.
+ * Copyright 2022 Nikolai Kotchetkov.
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
  * You may obtain a copy of the License at
@@ -13,8 +13,8 @@
 
 package com.motorro.rxlcemodel.rx
 
-import com.motorro.rxlcemodel.rx.LceState.Content
-import com.motorro.rxlcemodel.rx.LceState.Loading
+import com.motorro.rxlcemodel.lce.LceState.Content
+import com.motorro.rxlcemodel.lce.LceState.Loading
 import com.motorro.rxlcemodel.rx.entity.Entity
 import com.motorro.rxlcemodel.rx.entity.EntityValidator
 import com.nhaarman.mockitokotlin2.verify
