+++ rx/src/test/kotlin/com/motorro/rxlcemodel/rx/service/SyncDelegateCacheServiceTest.kt
@@ -1,5 +1,5 @@
 /*
- * Copyright 2019 Nikolai Kotchetkov.
+ * Copyright 2022 Nikolai Kotchetkov.
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
  * You may obtain a copy of the License at
@@ -13,8 +13,8 @@
 
 package com.motorro.rxlcemodel.rx.service
 
+import com.motorro.rxlcemodel.lce.LceState
 import com.motorro.rxlcemodel.rx.LceModel
-import com.motorro.rxlcemodel.rx.LceState
 import com.motorro.rxlcemodel.rx.entity.Entity
 import com.motorro.rxlcemodel.rx.entity.EntityValidator
 import com.nhaarman.mockitokotlin2.*
@@ -23,7 +23,6 @@ import io.reactivex.rxjava3.core.Single
 import org.junit.Before
 import org.junit.Test
 import java.util.*
-import kotlin.collections.HashMap
 
 class SyncDelegateCacheServiceTest {
     companion object {
