+++ sample/src/main/kotlin/com/motorro/rxlcemodel/sample/view/BaseLceModelFactory.kt
@@ -15,8 +15,8 @@ package com.motorro.rxlcemodel.sample.view
 
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.ViewModelProvider
+import com.motorro.rxlcemodel.lce.LceState
 import com.motorro.rxlcemodel.rx.LceModel
-import com.motorro.rxlcemodel.rx.LceState
 import com.motorro.rxlcemodel.rx.service.CacheService
 import com.motorro.rxlcemodel.rx.service.NetService
 import com.motorro.rxlcemodel.rx.service.ServiceSet
