+++ viewmodel/src/test/kotlin/com/motorro/rxlcemodel/view/LceStateViewTest.kt
@@ -15,7 +15,7 @@ package com.motorro.rxlcemodel.view
 
 import androidx.lifecycle.*
 import androidx.lifecycle.viewmodel.CreationExtras
-import com.motorro.rxlcemodel.rx.LceState
+import com.motorro.rxlcemodel.lce.LceState
 import com.motorro.rxlcemodel.viewmodel.BaseLceModel
 import com.nhaarman.mockitokotlin2.*
 import org.junit.Before
