+++ viewmodel/src/test/kotlin/com/motorro/rxlcemodel/viewmodel/BaseLceModelTest.kt
@@ -15,7 +15,7 @@ package com.motorro.rxlcemodel.viewmodel
 
 import androidx.arch.core.executor.testing.InstantTaskExecutorRule
 import androidx.lifecycle.Observer
-import com.motorro.rxlcemodel.rx.LceState
+import com.motorro.rxlcemodel.lce.LceState
 import com.nhaarman.mockitokotlin2.mock
 import com.nhaarman.mockitokotlin2.verify
 import io.reactivex.rxjava3.core.Completable
